# swagger.json
{"openapi":"3.0.1","info":{"title":"Credit Service API","description":"Апи для управления кредитами","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/employee/tariffs":{"post":{"tags":["Тарифы кредитов (Сотрудник)"],"summary":"Создать новый тариф кредита","operationId":"createTariff","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreditTariffRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditTariffResponse"}}}}}}},"/api/credits":{"get":{"tags":["Управление кредитами"],"summary":"Получить все кредиты","operationId":"getAllCredits","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreditResponse"}}}}}}},"post":{"tags":["Управление кредитами"],"summary":"Взять кредит","operationId":"takeCredit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TakeCreditRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/credits/{creditId}/repay":{"post":{"tags":["Управление кредитами"],"summary":"Погасить кредит","operationId":"repayCredit","parameters":[{"name":"creditId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepayCreditRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditPaymentResponse"}}}}}}},"/api/tariffs":{"get":{"tags":["Тарифы кредитов"],"summary":"Получить все доступные тарифы","operationId":"getAllTariffs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreditTariffResponse"}}}}}}}},"/api/tariffs/{id}":{"get":{"tags":["Тарифы кредитов"],"summary":"Получить тариф по ID","operationId":"getTariffById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditTariffResponse"}}}}}}},"/api/employee/tariffs/{id}":{"get":{"tags":["Тарифы кредитов (Сотрудник)"],"summary":"Получить тариф по ID","operationId":"getTariffById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditTariffResponse"}}}}}},"delete":{"tags":["Тарифы кредитов (Сотрудник)"],"summary":"Удалить тариф","operationId":"deleteTariff","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/credits/{id}":{"get":{"tags":["Управление кредитами"],"summary":"Получить информацию о кредите","operationId":"getCreditById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditResponse"}}}}}}},"/api/credits/{creditId}/statistics":{"get":{"tags":["Управление кредитами"],"summary":"Получить статистику по кредиту","operationId":"getCreditStatistics","parameters":[{"name":"creditId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreditStatisticsResponse"}}}}}}},"/api/credits/{creditId}/schedule":{"get":{"tags":["Управление кредитами"],"summary":"Получить график платежей","operationId":"getPaymentSchedule","parameters":[{"name":"creditId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentScheduleResponse"}}}}}}}},"/api/credits/{creditId}/payments":{"get":{"tags":["Управление кредитами"],"summary":"Получить историю платежей по кредиту","operationId":"getCreditPayments","parameters":[{"name":"creditId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreditPaymentResponse"}}}}}}}},"/api/credits/my":{"get":{"tags":["Управление кредитами"],"summary":"Получить мои кредиты","operationId":"getMyCreditsByToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreditResponse"}}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"CreditTariffRequest":{"required":["dueDate","interestRate","name"],"type":"object","properties":{"name":{"type":"string"},"interestRate":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"},"dueDate":{"type":"string","format":"date"},"isActive":{"type":"boolean"}}},"CreditTariffResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"interestRate":{"type":"number"},"dueDate":{"type":"string","format":"date"},"isActive":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"TakeCreditRequest":{"required":["amount","durationMonths","tariffId"],"type":"object","properties":{"tariffId":{"type":"integer","format":"int64"},"amount":{"minimum":1000.0,"exclusiveMinimum":false,"type":"number"},"durationMonths":{"type":"integer","format":"int32"}}},"CreditResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ownerId":{"type":"integer","format":"int64"},"tariffId":{"type":"integer","format":"int64"},"tariffName":{"type":"string"},"interestRate":{"type":"number"},"amount":{"type":"number"},"remainingAmount":{"type":"number"},"monthlyPayment":{"type":"number"},"durationMonths":{"type":"integer","format":"int32"},"remainingMonths":{"type":"integer","format":"int32"},"accumulatedPenalty":{"type":"number"},"overdueDays":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["ACTIVE","PAID_OFF","OVERDUE","CANCELLED"]},"issueDate":{"type":"string","format":"date-time"},"nextPaymentDate":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"RepayCreditRequest":{"required":["amount"],"type":"object","properties":{"amount":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"}}},"CreditPaymentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"creditId":{"type":"integer","format":"int64"},"amount":{"type":"number"},"paymentType":{"type":"string","enum":["MANUAL_REPAYMENT","AUTOMATIC_DAILY","EARLY_REPAYMENT","PENALTY"]},"paymentDate":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"CreditStatisticsResponse":{"type":"object","properties":{"creditId":{"type":"integer","format":"int64"},"originalAmount":{"type":"number"},"monthlyPayment":{"type":"number"},"durationMonths":{"type":"integer","format":"int32"},"totalToRepay":{"type":"number"},"totalInterest":{"type":"number"},"interestRate":{"type":"number"}}},"PaymentScheduleResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"creditId":{"type":"integer","format":"int64"},"monthNumber":{"type":"integer","format":"int32"},"paymentDate":{"type":"string","format":"date-time"},"totalPayment":{"type":"number"},"interestPayment":{"type":"number"},"principalPayment":{"type":"number"},"remainingBalance":{"type":"number"},"paid":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Введите JWT токен","scheme":"bearer","bearerFormat":"JWT"}}}}